Fix for #144: Unit tests should be bundled with component
 Also renamed utils and moved angular specific services to new folder ng-services
