+++ components/ng-services/pathresolver/pathresolver.ts
@@ -1,3 +1,3 @@
-///<reference path="../references.ts"/>
+///<reference path="../../references.ts"/>
 
 angular.module('jsonforms.pathresolver', []);
