+++ components/renderers/controls/array/array-renderer.spec.ts
@@ -1,5 +1,4 @@
-/// <reference path="../references.ts"/>
-
+/// <reference path="../../../references.ts"/>
 
 describe('Array renderer', () => {
