+++ components/renderers/layouts/group/GroupTest.ts
@@ -1,17 +1,17 @@
-/// <reference path="../references.ts"/>
+/// <reference path="../../../references.ts"/>
 
-describe('VerticalLayout', () => {
+describe('Group', () => {
 
     // load all necessary modules and templates
     beforeEach(module('jsonforms.form'));
-    beforeEach(module('jsonforms.renderers.layouts.vertical'));
+    beforeEach(module('jsonforms.renderers.layouts.group'));
     beforeEach(module('jsonforms.renderers.controls.string'));
 
     beforeEach(module('components/form/form.html'));
     beforeEach(module('components/renderers/layouts/layout.html'));
     beforeEach(module('components/renderers/controls/control.html'));
 
-    it("should not support labels", inject(($rootScope, $compile) => {
+    it("should support labels 2", inject(($rootScope, $compile) => {
         var scope = $rootScope.$new();
         scope.schema = {
             "properties": {
@@ -19,7 +19,8 @@ describe('VerticalLayout', () => {
             }
         };
         scope.uiSchema = {
-            "type": "VerticalLayout",
+            "type": "Group",
+            "label": "woot",
             "elements": [
                 {
                     "type": "Control",
@@ -33,6 +34,8 @@ describe('VerticalLayout', () => {
         var el = $compile('<jsonforms schema="schema" ui-schema="uiSchema" data="data"/>')(scope);
         scope.$digest();
         var div = el.find('div');
-        expect(div.hasClass('jsf-vertical-layout')).toBeTruthy();
+        expect(div.hasClass('jsf-group')).toBeTruthy();
     }));
 });
+
+
