+++ components/renderers/layouts/vertical/vertial.spec.ts
@@ -1,17 +1,17 @@
-/// <reference path="../references.ts"/>
+/// <reference path="../../../references.ts"/>
 
-describe('Group', () => {
+describe('VerticalLayout', () => {
 
     // load all necessary modules and templates
     beforeEach(module('jsonforms.form'));
-    beforeEach(module('jsonforms.renderers.layouts.group'));
+    beforeEach(module('jsonforms.renderers.layouts.vertical'));
     beforeEach(module('jsonforms.renderers.controls.string'));
 
     beforeEach(module('components/form/form.html'));
     beforeEach(module('components/renderers/layouts/layout.html'));
     beforeEach(module('components/renderers/controls/control.html'));
 
-    it("should support labels 2", inject(($rootScope, $compile) => {
+    it("should not support labels", inject(($rootScope, $compile) => {
         var scope = $rootScope.$new();
         scope.schema = {
             "properties": {
@@ -19,8 +19,7 @@ describe('Group', () => {
             }
         };
         scope.uiSchema = {
-            "type": "Group",
-            "label": "woot",
+            "type": "VerticalLayout",
             "elements": [
                 {
                     "type": "Control",
@@ -34,8 +33,6 @@ describe('Group', () => {
         var el = $compile('<jsonforms schema="schema" ui-schema="uiSchema" data="data"/>')(scope);
         scope.$digest();
         var div = el.find('div');
-        expect(div.hasClass('jsf-group')).toBeTruthy();
+        expect(div.hasClass('jsf-vertical-layout')).toBeTruthy();
     }));
 });
-
-
