+++ components/services/data/data-service.ts
@@ -1,4 +1,4 @@
-///<reference path="../references.ts"/>
+///<reference path="../../references.ts"/>
 
 declare module JSONForms {
