+++ components/services/rule/rule-service.d.ts
@@ -1,4 +1,4 @@
-///<reference path="../references.ts"/>
+///<reference path="../../references.ts"/>
 
 declare module JSONForms {
     export interface IRuleService extends IService {
