+++ components/services/rule/rule-service.spec.ts
@@ -1,4 +1,4 @@
-/// <reference path="../references.ts"/>
+/// <reference path="../../references.ts"/>
 
 describe('RuleServiceTest', () => {
