+++ components/services/rule/rule-service.ts
@@ -1,4 +1,4 @@
-///<reference path="../references.ts"/>
+///<reference path="../../references.ts"/>
 
 module JSONForms {
     export class RuleService implements IRuleService {
